Add support for openssh known_hosts files
Can read a file in the format of openssh's known_hosts, which allows identity verification of ssh hosts that you connect to.  Hosts can be added to a KnownHostStore object either programatically, or by parsing an openssh known_hosts file.
KnownHostStore can also be exported to text format, aloowing openssh or other compatible programs access to the hosts and their identities.  All features of the man 8 sshd secion on known_hosts are supported, except certificate authority
validation, which is not available in this library yet.