ENH: Implement SSH Jump as ProxyConnection
Rewrite some interfaces to allow this new type of ProxyConnection
Additionally, store and use information on proxy connections as a IConnectionInfo object "ProxyConnection" in the IConnectionInfo object. This allows the user to specify multiple proxies to connect through before making the final connection to the target (which will be done using a "DirectConnector" instance).

Update tests to reflect new interface