Implement support for rsa-sha2-256 key exchange.
There doesn't appear to be any way to select different exchange
methods depending on what the server supports for the same key, so
this commit allows IPrivateKeySource to expose multiple key
exchange methods for the same key and will try each one.

Tested on Ubuntu 18.04 and 22.04 with default SSH server config and
with only ssh-rsa and rsa-sha2-256 key exchange enabled.