Improve ShellStream
* Added SendWindowChangeRequest to ShellStream

* Added a new constructor to ShellStream

When creating ShellStream, the channel will be immediately opened in the constructor of ShellStream. At this time, the calling module may not have subscribed to ShellStream's DataReceived event, which may result in the loss of a small piece of data immediately after connection.
You can use ShellStream's new constructor to solve this problem